% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiragana2katakana.R
\name{hiragana2katakana}
\alias{hiragana2katakana}
\title{Convert hiragana to katakana}
\usage{
hiragana2katakana(x)
}
\arguments{
\item{x}{a string (contains hiragana)}
}
\value{
a string where all hiraganas are substituted into katakanas
}
\description{
Convert hiragana to katakana
}
