% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana2phone.R
\name{kana2phone}
\alias{kana2phone}
\title{Convert a kana string to a phone vector}
\usage{
kana2phone(str)
}
\arguments{
\item{str}{a katakana string}
}
\description{
Convert a kana string to a phone vector
}
