% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiragana2katakana.R
\name{katakana2hiragana}
\alias{katakana2hiragana}
\title{Convert katakana to hiragana}
\usage{
katakana2hiragana(x)
}
\arguments{
\item{x}{a string (contains katakana)}
}
\value{
a string where all katananas are substituted into hiraganas
}
\description{
Convert katakana to hiragana
}
