% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman2katakana.R
\name{roman2katakana}
\alias{roman2katakana}
\title{Convert romaji to katakana}
\usage{
roman2katakana(rom)
}
\arguments{
\item{rom}{a string (romaji)}
}
\value{
a string where all romaji are substituted into katakanas
}
\description{
Convert romaji to katakana
}
